requirements.txt
# Requirements per il sistema di classificazione avanzato Humanitas
# API Server e componenti avanzati

# Core dependencies (già presenti)
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0
torch>=1.9.0
transformers>=4.20.0
sentence-transformers>=2.2.0
hdbscan>=0.8.28
mysql-connector-python>=8.0.28
PyYAML>=5.4.1
requests>=2.25.1

# Database
pymongo>=4.5.0

# Web Framework
flask>=2.3.0

# ML Model persistence
joblib>=1.3.0

# HTTP utilities
urllib3>=2.0.0

# Nuove dipendenze per funzionalità avanzate

# FastAPI e server
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
pydantic>=2.0.0
python-multipart>=0.0.6

# Ottimizzazione e machine learning avanzato
optuna>=3.3.0
hyperopt>=0.2.7
bayesian-optimization>=1.4.0

# Visualizzazione e reporting
matplotlib>=3.5.0
seaborn>=0.11.0
plotly>=5.15.0
dash>=2.11.0

# Database e cache
redis>=4.6.0
sqlalchemy>=2.0.0
alembic>=1.11.0

# Monitoring e logging
prometheus-client>=0.17.0
structlog>=23.1.0

# Testing e quality assurance
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx>=0.24.0  # Per testing FastAPI

# Utilità
click>=8.1.0  # CLI tools
rich>=13.4.0  # Output formattato
tqdm>=4.65.0  # Progress bars

# Security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# Background tasks
celery>=5.3.0
flower>=2.0.0  # Monitoring Celery

# Documentation
mkdocs>=1.5.0
mkdocs-material>=9.1.0

flask-cors>=3.0.10

# Topic Modeling (BERTopic)
bertopic>=0.16.0
umap-learn>=0.5.5
scipy>=1.10.0